pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "termivox"
version = "0.2.0"
description = "Voice Recognition Bridge for Linux - Speak naturally, control your system, type hands-free with visual shortcuts"
readme = "README.md"
authors = [
    {name = "Gerico", email = "gerico@jgwill.com"}
]
license = {text = "MIT"}
# license = "MIT"  # Use when setuptools >= 77.0.0
keywords = ["voice-recognition", "speech-to-text", "linux", "vosk", "accessibility", "hands-free", "dictation", "shortcuts"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Desktop Environment",
    "Topic :: Adaptive Technologies",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
]
requires-python = ">=3.8"
dependencies = [
    "vosk>=0.3.50",
    "pyaudio>=0.2.14",
    "numpy>=1.26.0,<2.0",
    "pynput>=1.7.7",
    "pystray>=0.19.5",
    "Pillow>=11.0.0,<13.0",
    "speechrecognition>=3.14.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/Gerico1007/termivox"
Repository = "https://github.com/Gerico1007/termivox"
Issues = "https://github.com/Gerico1007/termivox/issues"

[project.scripts]
termivox = "termivox.main:main"
termivox-test = "termivox.test_voice_script:main"
termivox-download-model = "termivox.download_model:main"

[tool.setuptools]
package-dir = {"termivox" = "src"}

[tool.setuptools.package-data]
termivox = ["../config/*.json"]
